mcp: refactors pkg for easy maintainability (#2928)
